kafka_event_aggregator::config

Struct AggregatorConfig

pub struct AggregatorConfig {

    pub input_topic: String,
    pub output_topic: String,
    pub frame_queue_poll_interval_ms: Option<u64>,
    pub reference_time_tolerance_ns: Option<u64>,
    pub max_events_per_message: Option<usize>,
    pub expiry_offset_ms: Option<u64>,
    pub max_queued_frames: Option<usize>,
    pub sort_events_by_tof: Option<bool>,
    pub read_last_message_timeout_ms: Option<u64>,
    pub source_name: Option<String>,
    pub metrics_bind_addr: String,
    pub kafka_producer: HashMap<String, String>,
    pub kafka_consumer: HashMap<String, String>,

}

Global configuration flags for kafka_event_aggregator

Fields
input_topic: String

Input Kafka topic to read from (usually ends with _rawEvents)

output_topic: String

Output Kafka topic to write to (usually ends with _events)

frame_queue_poll_interval_ms: Option<u64>

How often to look for expired frames to send them to Kafka.

reference_time_tolerance_ns: Option<u64>

Reference-time comparison tolerance.

If two frames have equal reference times to within this tolerance, they are considered the same frame. The reference time of the frame is determined by the first message which arrives.

max_events_per_message: Option<usize>

Maximum number of neutron events to emit in each output Kafka message. Larger messages are somewhat more efficient, however the maximum resulting message size must fit within max.message.bytes on the Kafka broker.

expiry_offset_ms: Option<u64>

When a new frame arrives, it’s expiry time will be set to the current time plus this offset. All events for this frame are assumed to be consumed, in kafka_event_aggregator’s kafka producer, within this time.

Note that this may be affected by Kafka consumer batching and must account for the fact that messages from the same frame may be delivered in different Kafka batches.

max_queued_frames: Option<usize>

A maximum number of queued frames allowed to accumulate in this program. If there are more than this many frames, the oldest frames are emitted, even if they have not yet reached their expiry time.

sort_events_by_tof: Option<bool>

Whether to sort neutron events by time-of-flight before emitting them.

The input events are generally ToF ordered, however this process concatenates multiple events from multiple detectors, so the per-detector order is lost. This flag restores the ordering.

read_last_message_timeout_ms: Option<u64>

Timeout on reading the last message(s) emitted to the _events stream, to get the most recent message ID at startup. This configuration parameter is only used at startup.

source_name: Option<String>

String to insert into the source_name field of output pu00 and ev44 messages.

metrics_bind_addr: String

IP and port on which to bind the metrics server. Example: 127.0.0.1:8484

kafka_producer: HashMap<String, String>

Map of Kafka producer configuration properties. Values should be provided as strings. All properties are passed through to librdkafka.

kafka_consumer: HashMap<String, String>

Map of Kafka consumer configuration properties. Values should be provided as strings. All properties are passed through to librdkafka.

Implementations

impl AggregatorConfig

pub fn input_topic(&self) -> &str

pub fn output_topic(&self) -> &str

pub fn frame_queue_poll_interval_ms(&self) -> u64

pub fn reference_time_tolerance_ns(&self) -> u64

pub fn max_events_per_message(&self) -> usize

pub fn expiry_offset_ms(&self) -> u64

pub fn max_queued_frames(&self) -> usize

pub fn sort_events_by_tof(&self) -> bool

pub fn read_last_message_timeout_ms(&self) -> u64

pub fn source_name(&self) -> &str

pub fn metrics_bind_addr(&self) -> &str

pub fn kafka_producer_settings(&self) -> &HashMap<String, String>

pub fn kafka_consumer_settings(&self) -> &HashMap<String, String>